ntlmssp: get rid of "cast to pointer from integer of different size" warning on 64 bit
Not sure if this check is actually necessary, though:
the second should be sufficient.

Michael

+++ source/auth/ntlmssp/ntlmssp_parse.c
@@ -241,7 +241,7 @@ bool msrpc_parse(TALLOC_CTX *mem_ctx,
 					ret = false;
 					goto cleanup;
 				}
-				if (blob->data + ptr < (uint8_t *)ptr ||
+				if (blob->data + ptr < (uint8_t *)(uintptr_t)ptr ||
 						blob->data + ptr < blob->data) {
 					ret = false;
 					goto cleanup;
@@ -276,7 +276,7 @@ bool msrpc_parse(TALLOC_CTX *mem_ctx,
 					goto cleanup;
 				}
 
-				if (blob->data + ptr < (uint8_t *)ptr ||
+				if (blob->data + ptr < (uint8_t *)(uintptr_t)ptr ||
 						blob->data + ptr < blob->data) {
 					ret = false;
 					goto cleanup;
@@ -311,7 +311,7 @@ bool msrpc_parse(TALLOC_CTX *mem_ctx,
 					goto cleanup;
 				}
 
-				if (blob->data + ptr < (uint8_t *)ptr ||
+				if (blob->data + ptr < (uint8_t *)(uintptr_t)ptr ||
 						blob->data + ptr < blob->data) {
 					ret = false;
 					goto cleanup;
